requirements.txt
selenium
webdriver-manager
pandas
psycopg2-binary
selenium-stealth
sqlalchemy